Extract out ProductForm
